:recycle: Refactor: social-client 모듈 구축 완료
<footer>
- 관련: #134
