+++ application/external-api/src/main/java/com/namo/spring/application/external/domain/user/application/UserFacade.java
@@ -46,13 +46,13 @@ import com.namo.spring.application.external.domain.user.domain.constant.UserStat
 import com.namo.spring.application.external.domain.user.ui.dto.UserRequest;
 import com.namo.spring.application.external.domain.user.ui.dto.UserResponse;
 import com.namo.spring.application.external.global.common.constant.FilePath;
-import com.namo.spring.client.social.apple.AppleAuthClient;
-import com.namo.spring.client.social.apple.AppleProperties;
-import com.namo.spring.client.social.apple.AppleResponse;
-import com.namo.spring.client.social.apple.AppleResponseConverter;
-import com.namo.spring.client.social.apple.AppleUtils;
-import com.namo.spring.client.social.kakao.KakaoAuthClient;
-import com.namo.spring.client.social.naver.NaverAuthClient;
+import com.namo.spring.client.social.apple.client.AppleAuthClient;
+import com.namo.spring.client.social.common.properties.AppleProperties;
+import com.namo.spring.client.social.apple.dto.AppleResponse;
+import com.namo.spring.client.social.apple.dto.AppleResponseConverter;
+import com.namo.spring.client.social.common.utils.AppleUtils;
+import com.namo.spring.client.social.kakao.client.KakaoAuthClient;
+import com.namo.spring.client.social.naver.client.NaverAuthClient;
 import com.namo.spring.application.external.global.utils.FileUtils;
 import com.namo.spring.application.external.global.utils.JwtUtils;
 import com.namo.spring.application.external.global.utils.SocialUtils;
