--- application/external-api/src/main/java/com/namo/spring/application/external/global/config/properties/FeignConfig.java
@@ -1,9 +1,0 @@
-package com.namo.spring.application.external.global.config.properties;
-
-import org.springframework.cloud.openfeign.EnableFeignClients;
-import org.springframework.context.annotation.Configuration;
-
-@Configuration
-@EnableFeignClients("com.namo.spring.application.external.global.feignclient")
-public class FeignConfig {
-}
