--- application/external-api/src/main/java/com/namo/spring/application/external/global/feignclient/naver/NaverProperties.java
@@ -1,16 +1,0 @@
-package com.namo.spring.application.external.global.feignclient.naver;
-
-import org.springframework.boot.context.properties.ConfigurationProperties;
-import org.springframework.stereotype.Component;
-
-import lombok.Getter;
-import lombok.Setter;
-
-@Component
-@ConfigurationProperties(prefix = "spring.security.oauth1.client.registration.naver")
-@Getter
-@Setter
-public class NaverProperties {
-	private String clientId;
-	private String clientSecret;
-}
