--- clients/social-client/src/main/java/com/namo/spring/client/social/Main.java
@@ -1,7 +1,0 @@
-package com.namo.spring.client.social;
-
-public class Main {
-	public static void main(String[] args) {
-		System.out.println("Hello world!");
-	}
-}
