+++ clients/social-client/src/main/java/com/namo/spring/client/social/apple/config/AppleFeignConfiguration.java
@@ -1,0 +1,20 @@
+package com.namo.spring.client.social.apple.config;
+
+import org.springframework.context.annotation.Bean;
+
+import com.namo.spring.client.social.apple.decoder.AppleFeignClientExceptionDecoder;
+
+import feign.Logger;
+import feign.codec.ErrorDecoder;
+
+public class AppleFeignConfiguration {
+	@Bean
+	ErrorDecoder errorDecoder() {
+		return new AppleFeignClientExceptionDecoder();
+	}
+
+	@Bean
+	Logger.Level feignLoggerLevel() {
+		return Logger.Level.FULL;
+	}
+}
