+++ clients/social-client/src/main/java/com/namo/spring/client/social/common/FeignConfig.java
@@ -1,0 +1,9 @@
+package com.namo.spring.client.social.common;
+
+import org.springframework.cloud.openfeign.EnableFeignClients;
+import org.springframework.context.annotation.Configuration;
+
+@Configuration
+@EnableFeignClients(basePackages = "com.namo.spring.client.social")
+public class FeignConfig {
+}
