+++ clients/social-client/src/main/java/com/namo/spring/client/social/naver/client/NaverAuthClient.java
@@ -1,0 +1,54 @@
+package com.namo.spring.client.social.naver.client;
+
+import java.io.UnsupportedEncodingException;
+import java.net.URLEncoder;
+
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
+import org.springframework.stereotype.Component;
+
+import com.namo.spring.client.social.common.properties.NaverProperties;
+import com.namo.spring.client.social.naver.dto.NaverResponse;
+import com.namo.spring.client.social.naver.api.NaverAuthApi;
+
+import lombok.RequiredArgsConstructor;
+
+@Component
+@RequiredArgsConstructor
+public class NaverAuthClient {
+	private final Logger logger = LoggerFactory.getLogger(NaverAuthClient.class);
+	private final NaverAuthApi naverAuthApi;
+	private final NaverProperties naverProperties;
+
+	public String getAccessToken(String refreshToken) {
+		logger.debug("naver refreshToken : " + refreshToken);
+		NaverResponse.GetAccessToken getAccessToken = naverAuthApi.getAccessToken(
+			"refresh_token",
+			naverProperties.getClientId(),
+			naverProperties.getClientSecret(),
+			refreshToken
+		);
+		return getAccessToken.getAccessToken();
+	}
+
+	public void unlinkNaver(String accessToken) {
+		try { //accessToken URL인코딩
+			accessToken = URLEncoder.encode(accessToken, "UTF-8");
+		} catch (UnsupportedEncodingException e) {
+			e.printStackTrace();
+		}
+		logger.debug("clientId = {}, clientSecret = {}", naverProperties.getClientId(),
+			naverProperties.getClientSecret());
+
+		NaverResponse.UnlinkDto dto = naverAuthApi.unlinkNaver(
+			"delete",
+			naverProperties.getClientId(),
+			naverProperties.getClientSecret(),
+			accessToken,
+			"NAVER"
+		);
+
+		logger.debug("네이버 탈퇴 res : {}, {}", dto.getAccessToken(), dto.getResult());
+	}
+
+}
