+++ clients/social-client/src/main/java/com/namo/spring/client/social/naver/dto/NaverResponse.java
@@ -1,0 +1,33 @@
+package com.namo.spring.client.social.naver.dto;
+
+import com.fasterxml.jackson.annotation.JsonProperty;
+
+import lombok.Getter;
+import lombok.Setter;
+
+public class NaverResponse {
+	public NaverResponse() {
+		throw new IllegalStateException("Utility class.");
+	}
+
+	@Getter
+	public static class GetAccessToken {
+		@JsonProperty("access_token")
+		private String accessToken;
+	}
+
+	@Getter
+	@Setter
+	public static class UnlinkDto {
+		@JsonProperty("access_token")
+		private String accessToken;
+		private String result;
+	}
+
+	@Getter
+	public static class Availability {
+		private String resultcode;
+		private String message;
+	}
+}
+
